requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
rich>=13.0.0
tqdm>=4.65.0
